+++ Facebook.Unity/Results/ISoftKeyboardOpenResult.cs
@@ -1,0 +1,27 @@
+﻿/*As with any software that integrates with the Facebook platform, your use of
+ * this software is subject to the Facebook Developer Principles and Policies
+ * [http://developers.facebook.com/policy/]. This copyright notice shall be
+ * included in all copies or substantial portions of the software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+ * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+namespace Facebook.Unity
+{
+    /// <summary>
+    /// The result of a soft keyboard open
+    /// </summary>
+    public interface ISoftKeyboardOpenResult : IResult
+    {
+        /// <summary>
+        /// Gets the soft keyboard open result.
+        /// </summary>
+        /// <value>A string that specifies if the soft keyboard has open action was successful.</value>
+        string Success { get; }
+    }
+}