+++ Facebook.Unity/Windows/IWindowsFacebook.cs
@@ -23,5 +23,7 @@ namespace Facebook.Unity.Windows
     internal interface IWindowsFacebook : IPayFacebook, IFacebook
     {
         void SetVirtualGamepadLayout(string layout, FacebookDelegate<IVirtualGamepadLayoutResult> callback);
+
+        void SetSoftKeyboardOpen(bool open, FacebookDelegate<ISoftKeyboardOpenResult> callback);
     }
 }