+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [14.1.0]
 
+### Added
+- Windows platform methods
+    - FB.Windows.SetSoftKeyboardOpen
+
 ### Fixed
 - Issue with new Unity Input System and UnitySDK Example.
 - Issue with misconfiguration of each DLL when Unity import the package.